(working) move table to new component
